Some more explicit palette type handling
Original commit message from CVS:
Some more explicit palette type handling

+++ sys/v4l/gstv4lsrc.h
@@ -74,7 +74,7 @@ struct _GstV4lSrc {
   /* caching values */
   gint width;
   gint height;
-  gint palette;
+  guint16 palette;
 };
 
 struct _GstV4lSrcClass {
